+++ cloudpilot-ionic/src/app/model/DeviceOrientation.ts
@@ -1,0 +1,5 @@
+export enum DeviceOrientation {
+    protrait = 'portrait',
+    landscape90 = 'landscape90',
+    landscape270 = 'landscape270',
+}